adquando ao padraão
